refactor(timeline): セル描画を行クラスへ統合
