+++ lib/presentation/features/timeline/timeline_group_event_row.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:memora/presentation/features/group/group_event_edit_modal.dart';
-import 'package:memora/presentation/features/timeline/group_event_cell.dart';
 import 'package:memora/presentation/features/timeline/timeline_row_definition.dart';
 
 class TimelineGroupEventRow extends TimelineRowDefinition {
@@ -59,5 +58,41 @@ class TimelineGroupEventRow extends TimelineRowDefinition {
         },
       );
     };
+  }
+}
+
+class GroupEventCell extends StatelessWidget {
+  const GroupEventCell({
+    super.key,
+    required this.memo,
+    required this.availableHeight,
+    required this.availableWidth,
+  });
+
+  final String memo;
+  final double availableHeight;
+  final double availableWidth;
+
+  @override
+  Widget build(BuildContext context) {
+    final trimmedMemo = memo.trim();
+    if (trimmedMemo.isEmpty) {
+      return const SizedBox.shrink();
+    }
+
+    final maxLines = (availableHeight / 20).floor().clamp(1, 20);
+
+    return SizedBox(
+      width: availableWidth,
+      height: availableHeight,
+      child: Padding(
+        padding: const EdgeInsets.all(8),
+        child: Text(
+          trimmedMemo,
+          maxLines: maxLines,
+          overflow: TextOverflow.ellipsis,
+        ),
+      ),
+    );
   }
 }
